Allow configuring motor nSLEEP pin on MK6
On the MK6 the original design was to use a jumper to enable/disable the motor driver. This was in anticipation of no-motor utilization. That seems to be rare, and for motor installations we need to be able to bring the motor to a complete stop, so we can now configure what pin sleep is connected to.