Controlled shutdown on low battery
We already refuse to boot up when the battery is already low at boot. But we haven't so far handled the battery getting depleted while the device is in operation.

This change adds a recurring check and if the battery level drops below a threshold, and initiates a controlled shutdown. For now the only actual action while shutting down is closing open valves.