Suppress warning about both switches engaged
Looks like we have some reversed switches on the current implementation of the hardware,
and we end up having the switches reporting 0 when engaged, which leads to this error
being sent to the server constantly somehow, yet the device works well.

This is not a fix, only a workaround for the time being, because the chicken-door device
hardware is to be redesigned soon anyway. We can fix things properly then.