Improve handling of incoming MQTT message processing timeout
When an incoming message handling times out, we shouldn't abort the FreeRTOS task, because it can cause problems like #363.

Instead of having a hard deadline for the task and killing it, let's just trust the timeouts to work properly and handle things that way. This removes the whole question of having to deal with deleted tasks.

Fixes #363.