Improve MQTT timeouts
Connection and network timeouts are now both 15s.

Previously we waited for connection for 30s, which seems overly long. We also bailed on publishing and subscriptions after only 5s, while the client's network timeout was set to 10s for some reason. Based on real-world data 5 seconds might be too little for some of the roundtrip messages to make it, especially in a field environment where network might not be ideal.